requirements.txt
azure-core==1.23.0
azure-cosmos==4.3.0b3
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
